add drawChart module and remove logic from react component
